chore: refactor manager.Run function
This is pure refactor of the manager.Run function to
break individual sections down into discreet functions
for easier reading and maintenance. This includes no
changes to functionality.
