+++ railgun/manager/vars.go
@@ -1,5 +1,7 @@
 package manager
 
+import "github.com/kong/kubernetes-ingress-controller/pkg/util"
+
 // -----------------------------------------------------------------------------
 // Controller Manager - Versioning Information
 // -----------------------------------------------------------------------------
@@ -20,3 +22,11 @@ var (
 	//       See: "go doc cmd/link" for details, and "../Dockerfile.railgun" for invocation via "go build".
 	Commit = "UNKNOWN"
 )
+
+// -----------------------------------------------------------------------------
+// Controller Manager - Configuration Vars
+// -----------------------------------------------------------------------------
+
+// alwaysEnabled is a convenience alias to indicate that a component is not configurable
+// and is always enabled in any configuration.
+var alwaysEnabled = util.EnablementStatusEnabled
